Feature: test big conv network and modify float16 to float32

+++ MCNN/MCNN Shared/models/protocols/NetworkModuleProtocol.swift
@@ -8,6 +8,6 @@
 import Foundation
 
 public protocol NetworkModuleProtocol {
-    typealias DataType = Float16
+    typealias DataType = Float32
     func forward(input: Tensor<DataType>) -> Tensor<DataType>
 }
